Ban UNSAFE_CAST in lvalues
Summary:
From discussion on D32992694.
```
HH\FIXME\UNSAFE_CAST<mixed, dict<int, string>>($m)[0] = "lvalue";
```
UNSAFE_CAST's behavior is counterintuitive for the situation shown in the test case. A normal function in that position would return a value distinct from $m that then gets updated as a dictionary, whereas in this case we expect the expression to be equivalent to `$m[0] = "lvalue"`. To avoid this ambiguity, this diff bans UNSAFE_CAST expressions in lvalue expressions.

Reviewed By: enetsee

Differential Revision: D33000087

fbshipit-source-id: 74da85b9a9fbc0ccdeb9b02a82b859c72f48facb