Sys_utils.ncores
Summary:
We already have `Sys_utils.nproc` which returns the number of processors, e.g. on a machine with 18 cores and hyperthreading it will return the number 36.

This is a new function which returns the number of cores.

I've written this because I think that ncores is a better guide to how many workers are sensible to use, than nproc.

Differential Revision: D32986569

fbshipit-source-id: 80bc05a6152c81a2fe1b71d145d8569132c6be8f