Fix cursor position in autocomplete tests
Summary:
Autocomplete tests expect a file without the `AUTO332` marker (this represents the cursor), and a position.

We were passing in a file with an `AUTO332` marker already inserted, plus an incorrect position. This only worked because the the second AUTO332 token 'wins' (an implementation detail), and the incorrect position came first.

As a result, code completion would see files like this:

```
// Note that we're one line and one column too early.
function testTypecheck():AUTO332 void {
  HH\LongName\ShortName\foAUTO332;
}
```

Reviewed By: ljw1004

Differential Revision: D32975780

fbshipit-source-id: 9e86cb64e930f62b1f9462818bc6ee5c30f56c54