Don't emit AssertRATLs for certain minstr bytecodes
Summary:
There's a number of minstr bytecodes which are marked as "reading" a
local (because they update the known base type), but they don't
actually read the local. This causes us to emit redundant AssertRATLs
which don't help at all.

Add a check to the AssertRAT emission logic to filter out these cases.

Differential Revision: D31346740

fbshipit-source-id: ec9b37e89e900f6f5007244ff950b2aaf30cf5e2