add a flag to cap worker memory after each decl
Summary:
HOW TO REVIEW THIS DIFF:

The diff has a lot of files to read. They're almost all just boilerplate. The only real meat is
1. add a new setting in serverLocalConfig.ml (plus remove the old one) and serverLocalConfigKnobs.ml
2. thread the option into typing_service_types.ml
3. thread the option through typing_check_service.ml

-------------

Currently we use the "worker memory cap" `max_typechecker_worker_memory_mb` to say that: at the end of Typing_check_service.process_file, if a worker has exceeded this cap, then the worker should terminate and return all unprocessed files to the pool.

Reviewed By: Wilfred

Differential Revision: D32908203

fbshipit-source-id: a0178da3b4c89f2846be18d5e6868d44a507b0b9