Break RuntimeOptions out of mangleUnitSha1
Summary:
This is part of a stack of diffs breaking `mangleUnitSha1()` apart. In native Facts, I use `mangleUnitSha1()` to mutate the cache keys for source files, and to determine where to store the Facts DB. I want to be able to look inside and see what options are causing us to hit a new cache key or store the DB in a new place.

I've created a new function, called `RuntimeOption::getUnitCacheBreakingOptions()`, which contains all the inputs to `mangleUnitSha1()` coming from `HPHP::RuntimeOption`. In D32976419, I'll call this function inside of Facts so I can log HHVM's parser flags and track down changes.

I am changing the order of the runtime options passed to `mangleUnitSha1()`, but this shouldn't cause any stability problems because `mangleUnitSha1()` is already guaranteed to behave differently between each build of HHVM.

Reviewed By: ricklavoie

Differential Revision: D32976394

fbshipit-source-id: cdc9dd5f860a7c90534bba7bc7def56278f02443