Optimise Levenshtein distance
Summary:
During error reporting we try to find variables that the user might have meant if a binding is undefined. To do so we use the Levenshtein distance.

In the existing implementation, we were computing all the distances for suitable binders in the environment against a single target. This is not efficient because Levenshtein distance will keep computing distances even if we have a upper bound (the distance for the previous candidate).

This diff optimises the implementation by taking the upper bound into account.

The previous implementation also computed every distance twice. I also optimised that away by remembering the best distance across the fold.

Reviewed By: Wilfred

Differential Revision: D32723716

fbshipit-source-id: fbb784a819a1d2343fc65de5c7f0279291fc328d