Avoid collecting tyvar variance if env tyvars is empty
Summary: Optimize the `set_tyvar_variance_i` function in the situation that the set of fresh type variables is empty: there is no need to traverse a type looking for positive and negative occurrences of type variables. For some type checking scenarios this can have a significant effect on performance e.g. where the types of parameters or locals are very large, such as shapes.

Reviewed By: madgen

Differential Revision: D32795194

fbshipit-source-id: bb126a80531195497c8c7f94258f55d82a6cff91