Parsing error representation
Summary:
This is the first in a series of diffs to refactor `Errors.mli` to use a concrete representation of errors ahead of changes to error handling.

This diff introduces a datatype for parsing errors and, for the moment, modifies the existing function to construct the new representation before adding.

Ultimately, the idea is to manipulate these datatypes in naming, typing, tast-checks etc, with side-effects moved to the edge of each.

To do this, we have to refactor a few types and sigs currently *all* defined in `Errors.ml` into their own modules; these types correspond to the `Message.t` and `Quickfix.t` we need to be able to generate from the new error representation.

Reviewed By: Wilfred, vsiles

Differential Revision: D31539912

fbshipit-source-id: 7e3e35575a8f2a2d71cb027656b898fbc0431428