Simplify `Phase_error` interface and update parsing errors to implement it
Summary:
We only need access to individual components a `client_error` (i.e. code, claim, reasons, quickfixes) for a _typing_ error. To avoid boilerplate, this diff simplifies the interface for phase types so that we have a single pattern match over errors vs. 4.

The diff also inlines the old side-effecting errors that were in `errors.ml` to directly use the new error and a side-effecting `add_parsing_error` function.

Reviewed By: Wilfred

Differential Revision: D32023214

fbshipit-source-id: a3e8c00d4520b56dbc9c5c756f45d4e5f0039066