Move `error_` to its own module
Summary:
To allow conversion from our `Phase_error`s to the representation used for error reporting and avoid cyclic dependency with `errors.ml`, we need to move `error_` into and the `message` type alias to their own modules.

This diff moves the types along with (most of) the functions that operate on them. The notable exception to this is `to_string` which relies on mutable state inside `errors.ml`

The naming of `error_` is a little jarring but, unfortunately, the Jane Street core library exposes a module named `Error` too. ~~I went with `client_error` to indicate that this was the representation that is being sent by the server but I'm happy to bikeshed this!~~ The module is now named `user_error` to indicate that is provides user-readable information on which they must act to fix an error.

Reviewed By: Wilfred

Differential Revision: D32022574

fbshipit-source-id: 258a1b07e83d779ab7d3c19de876143a497b7496