Nast check error representation
Summary: Datatype for `Nast check` errors and refactoring of `Errors.ml` to construct this datatype before side-effecting `add_error`. As with the diff below this in the stack, this intermediate step is being taken before we use the datatype directly in naming, pushing side-effects to the edge of that code.

Reviewed By: Wilfred

Differential Revision: D31571674

fbshipit-source-id: 9a14d1667cbdb811ff8982e6b224783e5aec2ba6