Add back module arity check
Summary:
This was removed when moving to file level attributes ... because we aggressively strip file level attributes from the AST, so the check in `Attribute_nast_check` was doing nothing.

This seems like an OK place to add such a check, but I see arguments for putting this somewhere in `naming.ml`.

Reviewed By: Wilfred

Differential Revision: D32962855

fbshipit-source-id: 427c666ebf34805ed7d072e6e6e57fab8630479c