Rephrase typing errors using combinators
Summary:
Ahead of defunctionalizing error callbacks, this diff rephrases existing `typing_error_callback`s and `error_from_reason_callback`s into an opaque type rather than a type alias for a function and uses a set of combinators which will be directly translated into data constructors.

This diff makes it clearer that the role of a `Typing_error_callback` is to provide, at minimum, a default error code.

There is no guarantee that such a callback has a claim, and therefore primary position, rather claims are (typically) provided by `Error_from_reasons_callback`s.

`Error_from_reasons_callback`s are also used to carefully control which code, reasons and quickfixes are actually reported to the user.

The diff also moves the side-effecting application of callbacks to make it easier to defunctionalize these too.

Reviewed By: CatherineGasnier

Differential Revision: D31731106

fbshipit-source-id: e39b958584ac6ad4a0e0290aa408e5055809216f