Move BlobEncoder/BlobDecoder into hphp/util
Summary:
I want to use the BlobEncoder/BlobDecoder framework for more things
outside of hphp/runtime. Since this is a relatively generic utility,
it makes sense to put it in hphp/util/. The only problem is that it
implements serialization of some data structures which are specific to
hphp/runtime. Some of these data structures can be handled by adding
the appropriate serde member function to them. For the others, define
a customization point (BlobEncoderHelper) which allows for attach
serialization/deserialization logic to arbitrary types.

Reviewed By: edwinsmith

Differential Revision: D33053987

fbshipit-source-id: ea85bbeaa12547b918f81a408026a572b0460fde