Do not drop error codes when calling bad_decl_override
Summary:
Currently, the errors callback for a class inheritance error drops the individual code on the floor and always raises 4340. This is a problem for allowing a new error to be fixmeable because you are forced to make any 4340 (decl) error fixmeable, which is dangerous.

This diff threads through the error code in the callback, so the outer error reflects the code of the inner one.

Reviewed By: vassilmladenov

Differential Revision: D33081749

fbshipit-source-id: 5a138f6a630f30a86f7b1c3e56005629f25c42fd