Remove HHVM runtime options from the Facts schema hash
Summary:
See discussion in D33083826. I started mangling the Facts schema hash with runtime options so that future feature releases would be less likely to poison Facts saved state. This assumed:

1. New feature flags which might change Facts results will get included in the unit cache breaking options.
1. Feature flags will be kept consistent between CLI mode and server mode on the same host, so HHVM will reuse saved state when run in CLI mode or server mode.

It turns out (2) is explicitly not something we should expect. Given this, and given that none of the unit-cache breaking options today explicitly change Facts, I'm just going to remove this mangling.

Reviewed By: ricklavoie

Differential Revision: D33144638

fbshipit-source-id: 5a5ac9c64860febe1af3848325c2a0f6f26724c4