Internally use io::Error instead of write::Error
Summary:
Now that bytecode_printer is using io::Write as its stream we should just use io::Error instead of having everything return our custom error.

Unlike fmt::Error, io::Error has the ability to store foreign errors so in the rare cases where we want a specialized error we can just bury it in there.

A valid argument could be made that we should return anyhow::Error - but that would be a problem for code that has to vector through std::io routines (like a Display handler) so we'd need to handle having a buried error in io::Error anyway.

Since we define a `From<write::Error> for io::Error` much of this change is just adding `.into()` in places where we explicitly return a `write::Error`.

Reviewed By: shayne-fletcher

Differential Revision: D33061806

fbshipit-source-id: a91619c83f3bd62723f4e7038ae16f2301cd35c6