Change public interface for bytecode_printer to be byte-based
Summary:
Because of backward compatibility and FFI we end up using the output of bytecode_printer as a byte-stream rather than a UTF8-compliant stream.  As a result we really should treat the stream output as bytes and use io::Write instead of fmt::Write.

We could instead convert everything to UTF8 on input - but then we wouldn't match the existing mechanism used by hhvm (which allows raw non-utf8 in hhas files - and is actually tested in the hhvm tests). Also, what codepage would we use for conversion?

Reviewed By: shayne-fletcher

Differential Revision: D33061807

fbshipit-source-id: 5e011b8ba5fdbe2568e5839552e1f7b92ae14f7c