Correctly elide UNSAFE_CAST
Summary:
There are two circumstances where we emit the "wrong" code for `HH\FIXME\UNSAFE_CAST`.

(1) For expressions of the form `await HH\FIXME\UNSAFE_CAST<ty1,ty2>(foo(e1,...,en))` i.e. await a function call that's wrapped in a fixme. Currently, HHVM will fatal on this.

(2) For statements such as `HH\FIXME\UNSAFE_CAST<ty1,ty2>($lvar)[e1] = e2;` and similar, where `$lvar` is in an l-value position. The UNSAFE_CAST acts just as a Hack identity function would: the update "gets lost", because HackC doesn't get to special-case the use of a local in an l-value position.

This diff makes sure that both uses of `UNSAFE_CAST` are treated as if they simply aren't there. Arguably (2) should be outlawed, as its behaviour is different from normal functions.

In addition, there are a few places where we emit semantically equivalent but syntactically distinct bytecode. (Typically, it's in array access code where we make locals a special case.) I've added semantic tests for these too but don't know how to test the bytecode output - but I'm guessing that this is undesirable going forward anyhow.

To ensure that `UNSAFE_CAST` is *always* ignored, we elide it during the `closure_convert` phase. We already do a number of transformations in this phase that aren't strictly speaking related to closures.

[This is a replacement for the backed out diff D32992694 (https://github.com/facebook/hhvm/commit/0037794b2eefeb77104bc1c29d685f3c8368837f)]

Reviewed By: aorenste

Differential Revision: D33135376

fbshipit-source-id: a87a19f189c4d1ed6741f9e788f591fdd71bd44a