write_bytes macro
Summary:
io::Write is a string-based interface.  Even though the underlying stream is byte-based Display only gives you a stringy write interface (Display::fmt only gives you a fmt::Write - which only lets you write strings).

This adds a byte-based trait DisplayBytes which allows for byte-based writing to streams.

To make the trait more usable we also define a few helper macros which act ilke their non-byte counterparts but use DisplayBytes instead of Display:
- write_bytes! - works like write!
- writeln_bytes! - works like writeln!
- format_bytes! - works like format!

I originally tried using the format-bytes crate but it was deficient - it hasn't been updated in a while and had some unsurmountable bugs (the biggest being that `{{` printed `{{` instead of `{`).

The basic premise behind the write_bytes! macro is that it turns a static string into a sequence of `write_all()` and `write_bytes_fmt()` calls which actually write the data.

See the tests at the bottom of write_bytes-macro.rs to see how the macro converts input to output.

Reviewed By: shayne-fletcher

Differential Revision: D33062576

fbshipit-source-id: b403382f3a60649635f207036180108ee45a7ccb