introduce HackEventLogger.PerFileProfilingConfig.t containing all the profile_* configs
Summary:
There are multiple types which contain those `profile_*` fields, so we refactor those into HackEventLogger.PerFileProfilingConfig.t.

Why in HackEventLogger: by similarity with HackEventLogger.rollout_flags.

We also change HackEventLogger to take and store a PerFileProfilingConfig.t at `init` instead of a `profile_owner` and `profile_desc`, so we can later fine-tune how profiling info is logged into Scuba.

NB: In general it feels to me like HackEventLogger should have access to all options, which would avoid all the refs defined there. This is sort of a step in that direction.

Reviewed By: ljw1004

Differential Revision: D33061269

fbshipit-source-id: 03322bb6fee43e73199cfa53c6a2f51c98c5778c