Add a reference to the 'arena bump allocator into the emitter
Summary:
One potential refactor:

We have many functions in the code where we pass a reference to the arena allocator (the `bumpalo::Bump`) and the reference emitter object around as two arguments. We thought it might be cleaner to store the reference to the arena allocator in the emitter object. This way, if a function only passes the allocator through and doesn't use it, it doesn't need to do the extra work of threading two arguments along when it could simply thread one.

This diff is the first in a long series of diffs that attempts to execute on this refactor. This diff starts by inserting a reference to the allocator into the emitter object.

As this diff demonstrates, there are a couple of places that use the emitter that _do not_ use an arena allocator. In these places, I have created one for the purpose of the refactor. This could cause us to consider whether this is the correct refactor or whether the current system is good as is.

Reviewed By: shayne-fletcher

Differential Revision: D32898517

fbshipit-source-id: 38c44b1fec44ef9650598639cc1c5c0da3147c20