Remove alloc from toplevel emit_* and rewrite_* functions
Summary:
Continuation of refactor from D32898517.

In this diff, we continue refactoring the allocator argument from functions after we stuffed a reference to the allocator in the emitter.

This diff starts with the toplevel emit_* and rewrite_* functions. Instead of refactoring the whole tree of functions, we are going to do this piecemeal since we can pass `emitter.alloc` to function that haven't been refactored yet.

Reviewed By: shayne-fletcher

Differential Revision: D32899899

fbshipit-source-id: 3724413f458ea2cf78838a3ca514a7ec534b4886