Refactor emit_memoize_method
Summary:
Continue the refactor with `emit_memoize_method.rs`. This file is where I first noticed that we have a reference to the allocator in the env.rs object.

Moving forward, I assume that the allocator is the same as the one in the emitter and will attempt to simplify and refactor the env later in the stack.

Differential Revision: D32905495

fbshipit-source-id: af0c70d4de968efe7ab98c652536e9bbe85bc82c