Refactor ast_constant_folder
Summary:
Continue the refactor with `ast_constant_folder.rs`

This is one of the files that sees a significant reduction in complexity due to the fact that we had this `'local_arena` lifetime running around. I'm not quite sure what the deal with that was so I attempted to replace that with `'arena` which seems to work. Hopefully that's better overall.

Differential Revision: D32938321

fbshipit-source-id: 4c2661ddc79c9be45b452e38f829eff202d0db49