Push like through type constructors
Summary:
It is sound, and useful, to implement subtyping such as `(~int,string) <: ~(int,string)`. For example, consider
```
// For whatever reason, we need to pessimise the result
function get():~int { ... }

function makeTuple():~(int,string) {
  return tuple(get(), 'a');
}
```
In this diff, we implement the following rule in the case of built-in covariant types: tuples, shapes, `vec`, `dict`, `Awaitable`, and `keyset`. Here is the rule for tuples: it generalises in the obvious way to the other types.
```
t1 <:D dynamic ... tn <:D dynamic
---------------------------------------------
(~t1,...,~tn) <: ~(t1,...,tn)
```
Note the need for the component types to be (semantic) subtypes of `dynamic`.

The implementation is in two parts:

(1) Subtyping is extended to simplify the subtyping assertion `t <: ~(t1,...,tn)` to `t <: (~t1,...,~tn) & t1 <:D dynamic & ... & tn <:D dynamic`
(2) Bidirectional checking is extended so that for an expected type of `~(t1,...,tn)` we push the likes through to the expected types of the components.

Future diff will extend this to covariant user-defined generic types.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D33255568

fbshipit-source-id: cd301d4110decdfaa45bf100c5d8fcb6ec7b0951