Clean up switch case type mismatch lints/errors
Summary:
Type mismatches between cases and expressions being scrutinised in a switch statement are implemented so that they can be lints or errors.

This is governed by configuration. This diff removes the configuration and makes it a lint indiscriminately.

Apart from general cleanup of an undesirable configuration, this diff prevents double computation. The implementation of this dual behaviour was so that the linter framework would do unnecessary typing working and the type checker would do linter work regardless which configuration we're in.

Reviewed By: vsiles

Differential Revision: D33378982

fbshipit-source-id: d6e763389e7eaf07c484fd82229b64c1c65b3a44