Add switch exhaustiveness linter for non-enum types
Summary:
Adds exhaustiveness check in the form of a linter for non-enum types.

The linter is conservative and it looks for types that effectively have infinitely many values such as strings, ints, unions involving an infinite type, etc. If the switch statement involving these do not have a default case, then the switch statement is not exhaustive and a lint warning is generated.

The current implementation **should** be sound, but incomplete.

Reviewed By: vsiles

Differential Revision: D33380512

fbshipit-source-id: c4b9299b36e80c96f060a36fa654b03390dc6bba