hh --config memtrace_dir=<DIR>
Summary:
This diff adds memtrace for serverMain and for workers.

## Design considerations

The default way to use Memtrace is by setting the $MEMTRACE environment variable. You can't mutate an environment variable, so if we wanted to go down this path then we'd need to set it before `Daemon.spawn` and `Worker.ml:fork`.

Memtrace is strictly one file per process, and indeed it deletes the previous trace file if one exists. This is a problem for our multiworkers.

We normally like to start memtracing as soon as possible, sometimes even in the first line of the program, to capture every single memory allocation. (On the other hand, this is useless! We don't care about the trivial allocations encountered in the first small bits of startup. We only care about the bulk of the allocations).

We might want to turn on memtrace for every single process - for serverMain, for monitor, for each worker fork, for each worker prototype, for each scuba process. That added code complexity, and produced too many trace files that didn't seem useful in practice. I decided to simplify it down to just serverMain+workers.

I thought about merging the individual worker .ctf files at the end of a typecheck. But (1) if you did two typechecks in a row then we'd probably need the name of the merged ctf to reflect the recheck ID or the timestamp or similar, (2) what if you wanted to examine individual egregious workers? (3) the act of merging ctf files itself incurred a few GB of allocations, which showed up in the memtrace of serverMain. It's easier to leave merging to a standalone utility, unix philosophy, similar to the tools memtrace_hotspots and memtrace_dump that already ship with memtrace.

Memtrace has protection against fork: upon each memtrace, it checks whether the current PID is the same as the PID when it started. However, this protection landed recently, and our current supercaml+dune lack this protection. I don't think it matters. (1) Daemon.spawn uses Unix.create_process, which forks but then immediately execs, so the newly created process won't carry over any of its parent's memtracing. (2) Worker.ml prototype worker processes uses Fork.fork to create an actual child worker process, but the prototype itself hasn't yet started memtracing; that's left to the child. (3) The only other place we use fork is Daemon.fork, which is solely used by test, so I renamed it to make clear that it's test-only.

## What this diff does

All these considerations led to the design in this diff:
* Doesn't use $MEMTRACE environment variable with `Memtrace.trace_if_requested`. Instead use a more hack-conventional `--config memtrace_dir=<DIR>` and the more flexible `Memtrace.start_memtracing`.
* Only turn on memtrace at the few specific places we know we want it: (1) at the start of serverMain, (2) at the start of each worker fork
* Each worker fork traces into its own unique filename
* I wrote a standalone utility src/utils/memtrace:memtrace_merge to merge worker files, if desired:
```
./facebook/dune.sh build src/utils/memtrace/memtrace_merge.exe && \
   bin/memtrace_merge ~/mt/memtrace.worker.* > ~/mt/memtrace.workers.ctf
```

There's one known bug. I couldn't figure out how to produce proper timestamps in the merged worker trace. See TODO comment in code, which links to https://github.com/janestreet/memtrace/issues/14

Reviewed By: vsiles

Differential Revision: D32434129

fbshipit-source-id: c412bb163b35b00a1cf75b4ed2c43c50f6ea8ab5