Ignore `vendor/bin/` in open source builds of Hack
Summary:
Previously, `vendor/bin/example.hack` would be a symlink to
`vendor/foocorp/barlib/bin/example.hack`

As of composer 2.2.0, these are shell script wrappers instead of
symlinks, so they now raise type errors.

Always ignore as we aim for Hack to be usable with an empty .hhconfig, and
all projects will need this.

Reviewed By: viratyosin

Differential Revision: D33410895

fbshipit-source-id: 50ebfda82d74823f55c2839ef3407cc1566aed54