Use fuel in type printing
Summary:
The typechecker used to hang when a big type was encountered and needs to be printed out in an error message.

Also large types in error messages are effectively useless and hinders browseability of all errors.

This diff fixes it by introducing fuel to the decl and locl type printers. When we run out of fuel, we print `[truncated]` for all subsequent types. The overall type is also marked as abridged in the output string to prevent confusion.

The amount of fuel the printers have is governed by configuration. In `hhstc`, we use `--type-printer-fuel N` option and in `hh` we use `--config type_printer_fuel=N`. Sadly, to increase or decrease the fuel, we need to stop the typechecker first and restart with the new configuration. Hopefully, increasing fuel won't be a common workflow.

There are many possible choices for what amount of computation one fuel would correspond to, but this diff uses a simple scheme. Each recursive call to locl type or decl type printer is one fuel.

The default number of fuel is set to 100. That should comfortably accommodate very large types.

The main file that requires careful review is `typing_print.ml` which has fuel threaded throughout the file. The rest of the files is to thread the typechecker option for fuel to the printers.

All abridged type strings also contain instructions to increase fuel, i.e., `[This type was truncated. If you need more of the type, use hh stop; hh --config type_printer_fuel=N. At this time, N is 100.]`

Reviewed By: vsiles

Differential Revision: D33358668

fbshipit-source-id: fb180e3e25ee8a9840c860a4ee14dd2f1714a9c5