Don't use global mutable state for local variable completion
Summary:
Previously, we would update several mutable refs whilst inspecting the TAST, compute partial autocomplete results, and only process them later. This was error prone as it was easy to miss some refs.

Instead, compute local variable completion results as soon as we find the relevant point in the TAST, and don't use the mutable refs.

Reviewed By: hgoldstein

Differential Revision: D32661833

fbshipit-source-id: 356fad7bbdbdaaf69b005e70e7aba5666c96d47e