Don't use global mutable state for shape field completion
Summary:
Previously, we would update several mutable refs whilst inspecting the TAST, compute partial autocomplete results, and only process them later. This was error prone as it was easy to miss some refs.

Instead, compute shape completion results as soon as we find the relevant point in the TAST, and don't use the mutable refs.

Reviewed By: hgoldstein

Differential Revision: D32662086

fbshipit-source-id: dac62858775359a4476f5d79440d820c492be76a