Don't use global mutable state for XHP completion
Summary:
Previously, we would update several mutable refs whilst inspecting the TAST, compute partial autocomplete results, and only process them later. This was error prone as it was easy to miss some refs.

Instead, compute XHP completion results as soon as we find the relevant point in the TAST, and don't use the mutable refs.

Reviewed By: hgoldstein

Differential Revision: D32709592

fbshipit-source-id: e9143a85b69402ca29d47d6800f4cecdc727adcd