Move HHVM config parsing to utils
Summary:
In the diff above, I want to use be able to parse the hhvm config files (hdf and ini) as well as support ini and hdf cli arguments for hh_single_compile_cpp. Hack Native has some of the capability to do that in `hh_distc` which is under a `facebook` directory.

Instead of having a dependency on a `facebook/` dependency for our hh_single_compile ffi, we should move this capability out into utils so that we don't have hidden dependencies here.

Reviewed By: aorenste

Differential Revision: D33438987

fbshipit-source-id: cfd371f9066ddb8d184f1eb4f7b8df67d04efcc5