Turn on runtime local coeffect enforcement and add tests
Summary: To avoid the situation where someone commits a new WWW coeffect error not amended in D32268019 and hhvm enforces it before the hack release (they will not catch it in development) we only released changes in the typechecker. This diff opens it to runtime now that the typechecker changes are released and adds the hhvm test counterparts to the existing typechecker tests. We can also remove the logic around pure (we preserved hhvm checking for write_props in pure functions since it was already committed but now checks are enforced on all functions, not just pure).

Reviewed By: oulgen

Differential Revision: D32708938

fbshipit-source-id: 1d88a9ac1930505029d7249106bc8bd7b9d9fbfb