Change How We Handle Disk-Writing Errors While Logging
Summary: Currently we assert on success for fwrite() and fflush() in the cronologwriter.  It seems there are some machines for which the writes are failing and we don't really want to crash the binary in that case.  This adds a new flag that decides how to handle log entries which cannot be written to disk:  write them out to stderr or drop them.  This defaults to just dropping the errors.

Reviewed By: jthemphill

Differential Revision: D33467256

fbshipit-source-id: e5f8adb1a2cc3424974fa889bed9578e0f8ca403