Add extension
Summary:
The analysis now reports on intermediate points where a shape grows. Particularly, when it is extended with an assignment to one of its keys.

For example
```
$d = dict['a' => 42] // Allocation
$d['b'] = true; // Extension
```

The reason we can care about this is even the extension was by a dynamic key, we could use the allocated (or previously extendd) dict as a shape and cast to dict when it no longer agrees with shape constraints. This identifies more (albeit less impactful) targets.

Reviewed By: enetsee

Differential Revision: D32919686

fbshipit-source-id: b92a561bc7a1a88bb6db02d2ad9b3019b24708d0