Copy-on-write array assignments
Summary:
This implements copy-on-write assignment by creating an auxiliary variables on array assignments and recording the key on that variable and establishing a subset relationship with the previous entity.

For now, this makes the results even less interesting. We'll soon have extension points in addition to allocation points, then we'll see interesting shape types discovered again.

Reviewed By: enetsee

Differential Revision: D32919687

fbshipit-source-id: b197c8d20dbc086bf815ee1d7e63ff197606a357