Slap some CPGO on hphp_debug_caller_info
Summary:
We might as well assign a RuntimeStruct to this dict constructor so that we can produce struct-shaped arrays here.

When we const-fold the function, we now have to apply the selected layout to the result (as GetScalarArray always coerces the result back into a vanilla layout).

This change caused us to use RuntimeStruct during JIT time, which exposed a somewhat subtle (to me) bug. In getSrcKey(), we aim to detect cases where we're not in the VM and bail, using a "soft" VMRegAnchor. This mechanism almost works fine, but fails during JIT time in debug mode, because we protect the VM registers (see: vm-protect.cpp - thanks ricklavoie) and the destructor of a soft VMRegAnchor writes back to these registers. I modified "soft" VMRegAnchor to avoid the write-back in debug mode.

Reviewed By: ricklavoie

Differential Revision: D33443218

fbshipit-source-id: 9a664537eb84c0e9512ce5188a8e8eaf261f3380