Improve assignment to keyset/dict
Summary: If we have an assignment to a keyset `$k[] = i` or dict `$d[i] = exp`, then if `i` is not a subtype of arraykey, and we don't get a type error, `dynamic` must have been involved in the type of `i`. Whatever `i`'s type, we should set the key type of the result to to `arraykey`, since the runtime will enforce that the key is an arraykey, and we don't know if it is a string or an int. The previous version checked for `dynamic` explicitly, and wasn't robust wrt intersections. We now do the more straightforward thing and just check if the type is a subtype of arraykey.

Reviewed By: andrewjkennedy

Differential Revision: D33509924

fbshipit-source-id: 5f216f7395d75b12f3ef5a58c5bad36c78236cee