Add support for attributes on class constants
Summary:
Update Ast and parser to allow for attributes to be placed on class constants.

Currently, this is only taking care of class constants and not enum class constants. vsiles has separate work that is specific to enum class constants (D32675298)

Immediate motivation behind this is to allow a new `<<Exhaustive>>` attribute which can be put on constant dictionaries throughout Meta (D32934092 stack).

Reviewed By: vsiles

Differential Revision: D32955735

fbshipit-source-id: 0e01a19ee78434e19f25322691721b4cffeb855e