Rework canonicalization logic for DefLabels in refineTmps
Summary:
refineTmps keeps a visited set when canonicalizing to avoid
going into infinite loops. If we hit in the visited set, we ignore
that branch of the recursion. However, we only ever added to the set,
and didn't remove from it when unwinding the recursion. This isn't
correct. A DefLabel can be visited via multiple paths even in
non-ecursive situations, and just because we visited it once already
doesn't mean we can simply ignore it on all other paths.

As a result, we can end up ignoring paths which are actually relevant,
and calculate an incorrect canonicalization. One solution is to just
remove from the visited set when returning. An element should only be
in the set if it's part of the current call chain in the
recursion. However this causes us to visit many redundant paths which
we didn't before. As a result, compilation time can become
pathologically large.

Instead rework the canonicalization logic entirely. Attempting to
calculate the canonical value for a DefLabel on-demand has proven
difficult and bug prone. The problem is that to avoid exponential path
blowup, you need some kind of caching to avoid redundant
paths. However caching interacts poorly with DefLabels that are
mutually recursive.

Instead of calculating DefLabel canonical values on the fly,
pre-calculate them all in a separate pass ahead of time. During the
main analysis, all we have to do is consult our pre-calculated values
for the canonical value. All the recursion and visited logic just goes
away. Calculating canonical values for non-DefLabels is still trivial,
so nothing needs to be pre-calculated for those. The calculation is a
straight forward application of dataflow. First we build up
dependencies between DefLabels (a DefLabel depends on another if one
of its sources comes from that DefLabel). Then we use a worklist and
process each DefLabel (repeating until worklist is empty). For each
DefLabel, iterate over it's sources and see if all the sources
canonicalize to the same value. If the source is not a DefLabel, the
canonical value is itself. Otherwise we use the (potentially not
final) calculated canonical value for that source DefLabel. If all
sources canonicalize to the same value, we use that as the DefLabel's
canonical value. Otherwise we use the DefLabel itself. If we calculate
a new canonical value which is different than the previous, we put any
dependent DefLabels onto the worklist to process again. When
processing sources, if a source DefLabel does not yet have a value
(hasn't been processed yet), we just ignore it. This is similar to
dealing with a Bottom in normal dataflow. The dataflow algorithm takes
care of all the subtlty with mutually recursive DefLabels.

Reviewed By: ottoni

Differential Revision: D33503510

fbshipit-source-id: 127e0a5d75f843828f0b0b986b2152e66949ea01